[cake] Remove old compute method and old cache
